# docs/swagger.yaml
definitions:
  entities.Stock:
    properties:
      amount:
        type: integer
      categoryId:
        type: integer
      expireDate:
        type: string
      id:
        type: integer
      name:
        type: string
      userId:
        type: integer
    type: object
  entities.StockCategory:
    properties:
      id:
        type: integer
      name:
        type: string
      userId:
        type: integer
    type: object
  entities.User:
    properties:
      id:
        type: integer
      mailAddress:
        type: string
      name:
        type: string
    type: object
  entities.UserAuth:
    properties:
      authProvider:
        type: string
      uid:
        type: string
      userId:
        type: integer
    type: object
info:
  contact: {}
paths:
  /auth/signin:
    post:
      parameters:
      - description: UserAuth
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.UserAuth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Providerによるサインインを実施する。新規ユーザの場合、ユーザ登録処理を実施する。
      tags:
      - Auth
  /categories:
    get:
      parameters:
      - description: userId
        in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.StockCategory'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 条件に一致するCategoryを取得
      tags:
      - Category
    post:
      parameters:
      - description: Category
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.StockCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 登録したCategory
          schema:
            $ref: '#/definitions/entities.StockCategory'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Categoryを1件登録
      tags:
      - Category
  /categories/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: idで指定したCategoryを1件削除する
      tags:
      - Category
    put:
      parameters:
      - description: ID
        in: path
        name: id
        type: string
      - description: Category
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.StockCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: idで指定したCategoryを1件更新する
      tags:
      - Category
  /stocks:
    get:
      parameters:
      - description: userId
        in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Stock'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 条件に一致するStockを取得
      tags:
      - Stock
    post:
      parameters:
      - description: Stock
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.Stock'
      produces:
      - application/json
      responses:
        "200":
          description: 登録したStock
          schema:
            $ref: '#/definitions/entities.Stock'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Stockを1件登録
      tags:
      - Stock
  /stocks/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: idで指定したStockを1件削除する
      tags:
      - Stock
    put:
      parameters:
      - description: ID
        in: path
        name: id
        type: string
      - description: Stock
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.Stock'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: idで指定したStockを1件更新する
      tags:
      - Stock
  /user-auths:
    get:
      parameters:
      - description: userId
        in: query
        name: userId
        type: string
      - description: authProvider
        in: query
        name: authProvider
        type: string
      - description: uid
        in: query
        name: uid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.UserAuth'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: 検索条件に合うUserを取得
      tags:
      - User
  /users/{id}:
    put:
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: User
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: userIdに対応するユーザの情報を更新する
      tags:
      - User
swagger: "2.0"